Fix numpy 1.20 deprecation warnings

+++ hdbscan/flat.py
@@ -331,7 +331,7 @@ def approximate_predict_flat(clusterer,
         else:
             return labels, probabilities
 
-    labels = np.empty(points_to_predict.shape[0], dtype=np.int)
+    labels = np.empty(points_to_predict.shape[0], dtype=np.int32)
     probabilities = np.empty(points_to_predict.shape[0], dtype=np.float64)
 
     min_samples = clusterer.min_samples or clusterer.min_cluster_size
